requirements.txt
fastapi
uvicorn
requests
Pillow
torch
faiss-cpu
streamlit
git+https://github.com/openai/CLIP.git